export URL either to 'url' or 'opt_url'
The translator now will export Zotero's 'URL' field for all items.
(1) for items that are a 'book', 'book section' or 'thesis' it will
export the URL to 'opt_url'.  It will do likewise for an item of any
type which has a value for the field 'pages' (under the assumption
that these are primarily print-based resources).

For resources which do not have a value for the 'pages' field, the
URL is exported as with the stock Zotero BibTeX export in the 'url'
field.

The resaon for this is that some bibtex styles will include the URL if
it is in the .bib datafile.  But we mostly don't want this for
print resources.